Move CardMetadata types, GetCardMetadata, and their tests from cards.go/cards_test.go into dedicated card_metadata.go/card_metadata_test.go files.
